New function Export()
Export( .var ) is the counter part of Using( .var ).
It will update the previous definition of .var in the first ancestor scope declaring it.
If a previous declaration of .var is not found the parent scope will be used to declare a new .var.

This addition allows the following idiom :

.Defines = { '_DEBUG', 'SOMETHING=1' }
.CompilerOptions = ''

#if __WINDOWS__
ForEach( .name in .Defines )
{
   .CompilerOptions + ' /D "$name$"'
   Export( .CompilerOptions )
}
#endif

#if __LINUX__
ForEach( .name in .Defines )
{
   .CompilerOptions + ' -D$name$'
   Export( .CompilerOptions )
}
#endif

Print( "$CompilerOptions$" ) ; windows : ' /D "_DEBUG" /D "SOMETHING=1"'
                             ; linux   : ' /D_DEBUG -DSOMETHING=1'

Signed-off-by: poppolopoppo <frostiebek@gmail.com>